put full timeline in My LII view
