+++ peachjam/templates/peachjam/my/home.html
@@ -10,7 +10,7 @@
       <div class="col-lg-6">{% include "peachjam/my/_saved_searches.html" %}</div>
     {% endif %}
     {% if perms.peachjam.view_userfollowing %}
-      <div class="col-lg-12">{% include "peachjam/my/_recent_additions.html" %}</div>
+      <div class="col-lg-12" id="timeline">{% include "peachjam/my/_recent_additions.html" %}</div>
     {% endif %}
   </div>
 {% endblock %}
