+++ peachjam/templates/peachjam/user_following/list.html
@@ -36,10 +36,4 @@
       {% endfor %}
     </ul>
   </div>
-  {% if following_timeline %}
-    <h5 class="mt-4 mb-3" id="timeline">{% trans "What's new in topics you follow" %}</h5>
-    <ol class="list-timeline">
-      {% include "peachjam/user_following/_timeline.html" %}
-    </ol>
-  {% endif %}
 {% endblock %}
