+++ peachjam/urls/my.py
@@ -8,11 +8,13 @@ from peachjam.views import (
     FolderUpdateView,
     MyFrontpageView,
     MyHomeView,
+    MyTimelineView,
     UserFollowingListView,
 )
 
 urlpatterns = [
     path("", MyHomeView.as_view(), name="my_home"),
+    path("timeline", MyTimelineView.as_view(), name="my_timeline"),
     path("frontpage", MyFrontpageView.as_view(), name="my_frontpage"),
     path(
         "folders/",
